playwright/requirements.txt
playwright==1.44.0
flask
psycopg2-binary
